requirements.txt
# requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
python-dotenv==1.0.0
python-multipart
openpyxl
passlib[bcrypt]==1.7.4
bcrypt<4
python-jose[cryptography]==3.3.0
